cleanup dialogue manager
